MDEV-5228 Fix incorrect message on a failed attempt to revoke grants from a role
Ensure that when there is a failed attempt to revoke a grant
for a role, the error message indicates that the failure was
on a role instead of a user.

All new code of the whole pull request, including one or several files
that are either new files or modified ones, are contributed under the
BSD-new license. I am contributing on behalf of my employer Amazon Web
Services, Inc.